django_crop/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-cors-headers==3.13.0
fdfs-client-py==1.2.6
idna==3.4
mutagen==1.40.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2022.7.1
requests==2.31.0
sqlparse==0.4.3
urllib3==2.0.2
